BOL-35 chore: 충돌 해결 커밋
