+++ adapter-out/rdb/src/main/kotlin/com/yapp/bol/group/member/MemberRepository.kt
@@ -7,17 +7,3 @@ import org.springframework.stereotype.Repository
 internal interface MemberRepository : JpaRepository<MemberEntity, Long> {
     fun findByNickname(nickname: String): MemberEntity?
 }
-
-fun MemberEntity.toDomain(): Member = Member(
-    id = this.id,
-    userId = this.userId,
-    role = this.role,
-    nickname = this.nickname,
-)
-
-fun Member.toEntity(): MemberEntity = MemberEntity(
-    id = this.id,
-    userId = this.userId,
-    role = this.role,
-    nickname = this.nickname,
-)
